Add tests for plot
